services/booking-service/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
psycopg2-binary==2.9.9
redis==5.0.1
gunicorn==21.2.0
pytest==7.4.3

